requirements.txt
Django
djangorestframework
python